+++ src/math/llroundl.c
@@ -15,7 +15,7 @@ static struct l_i t[] = {
 int main(void)
 {
 	#pragma STDC FENV_ACCESS ON
-	int yi;
+	long long yi;
 	float d;
 	int e, i, err = 0;
 	struct l_i *p;
@@ -37,7 +37,7 @@ int main(void)
 			err++;
 		}
 		if (yi != p->i) {
-			printf("%s:%d: %s llroundl(%La) want %lld got %d\n",
+			printf("%s:%d: %s llroundl(%La) want %lld got %lld\n",
 				p->file, p->line, rstr(p->r), p->x, p->i, yi);
 			err++;
 		}
